Split out config view and clear
